+++ scripts/make_release.sh
@@ -1,5 +1,6 @@
 #!/usr/bin/env bash
-./scripts/uninstall.sh
-./scripts/clean_up.sh
-python setup.py register
-python setup.py sdist bdist_wheel upload
+#./scripts/uninstall.sh
+#./scripts/clean_up.sh
+#python setup.py register
+python setup.py sdist bdist_wheel
+twine upload dist/* --verbose